+++ .github/workflows/tests.yml
@@ -317,8 +317,6 @@ jobs:
     permissions:
       contents: 'read'
       id-token: 'write'
-      issues: write
-      pull-requests: write
 
     runs-on: ${{ matrix.os }}
 
@@ -412,5 +410,5 @@ jobs:
           SQLSERVER_PASS: '${{ steps.secrets.outputs.SQLSERVER_PASS }}'
           SQLSERVER_DB: '${{ steps.secrets.outputs.SQLSERVER_DB }}'
         if: "${{ matrix.node-version != 'v14.x' }}"
-        run: npx tap -c -t0 --disable-coverage --allow-empty-coverage examples/*/{mysql2,pg,tedious}.test{.cjs,.mjs,.ts} -o test_results.tap
+        run: npx tap -c -t0 --disable-coverage --allow-empty-coverage examples/*/{mysql2,pg,tedious}/*.test{.cjs,.mjs,.ts} -o test_results.tap
         timeout-minutes: 5
