+++ examples/typeorm/mysql2/connect.test.cjs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 const t = require('tap');
-const connector = require('./mysql2.cjs');
+const connector = require('./connect.cjs');
 
 t.test('mysql2 typeorm cjs', async t => {
   const { dataSource, close } = await connector.connect({
@@ -21,8 +21,8 @@ t.test('mysql2 typeorm cjs', async t => {
     username: process.env.MYSQL_IAM_USER,
     database: process.env.MYSQL_DB,
   });
-  const [{ now: time }] = await dataSource.manager.query('SELECT NOW() as now')
-  t.ok(time.getTime(), 'should have valid returned date object');
+  const [{ now }] = await dataSource.manager.query('SELECT NOW() as now')
+  t.ok(now.getTime(), 'should have valid returned date object');
   await close();
 });
