+++ examples/typeorm/mysql2/connect.test.ts
@@ -12,17 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-const t = require('tap');
-const connector = require('./tedious.cjs');
+import t from 'tap';
+import {connect} from './connect';
 
-t.test('tedious typeorm cjs', async t => {
-  const { dataSource, close } = await connector.connect({
-    instanceConnectionName: process.env.SQLSERVER_CONNECTION_NAME,
-    username: process.env.SQLSERVER_USER,
-    password: process.env.SQLSERVER_PASS,
-    database: process.env.SQLSERVER_DB,
+t.test('mysql typeorm ts', async t => {
+  const {dataSource, close} = await connect({
+    instanceConnectionName: process.env.MYSQL_IAM_CONNECTION_NAME,
+    username: process.env.MYSQL_IAM_USER,
+    database: process.env.MYSQL_DB,
   });
-  const [{ now: time }] = await dataSource.manager.query('SELECT GETUTCDATE() as now')
-  t.ok(time.getTime(), 'should have valid returned date object');
+  const [{now}] = await dataSource.manager.query('SELECT NOW() as now');
+  t.ok(now.getTime(), 'should have valid returned date object');
   await close();
 });
