+++ examples/typeorm/tedious/connect.test.ts
@@ -13,15 +13,16 @@
 // limitations under the License.
 
 import t from 'tap';
-import {connect} from './mysql2';
+import {connect} from './connect';
 
-t.test('mysql typeorm ts', async t => {
+t.test('tedious typeorm ts', async t => {
   const {dataSource, close} = await connect({
-    instanceConnectionName: process.env.MYSQL_IAM_CONNECTION_NAME,
-    username: process.env.MYSQL_IAM_USER,
-    database: process.env.MYSQL_DB,
+    instanceConnectionName: process.env.SQLSERVER_CONNECTION_NAME,
+    username: process.env.SQLSERVER_USER,
+    password: process.env.SQLSERVER_PASS,
+    database: process.env.SQLSERVER_DB,
   });
-  const [{now: time}] = await dataSource.manager.query('SELECT NOW() as now');
-  t.ok(time.getTime(), 'should have valid returned date object');
+  const [{now}] = await dataSource.manager.query('SELECT GETUTCDATE() as now');
+  t.ok(now.getTime(), 'should have valid returned date object');
   await close();
 });
